% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_address.R
\name{data_address}
\alias{data_address}
\title{Longitude and Latitude of a address/destination}
\usage{
data_address(address)
}
\arguments{
\item{address}{shall be a address/destination of class character}
}
\value{
Returns a data.frame of the name of the address/destination, longitude and latitude.
}
\description{
Return the longitude and latitude of a address/destination
}
\examples{
data_address("linkoping")
data_address("london")
}
