% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5InYourFace.R
\docType{package}
\name{Lab5InYourFace}
\alias{Lab5InYourFace}
\alias{Lab5InYourFace-package}
\title{Lab5InYourFace: Download data or Plot a address/destination!}
\description{
This package use API from googlemaps to download Longitude and Latitude of a address/destination so you can plot it!
}
\section{The packages functions}{

\describe{
  \item{data_address}{Data.frame of longitude and latitude for the address/destination}
   \item{plot_address}{Plot the address/destination}
}
}

\author{
Albin Vasterlund and Eric Herwin
}
