% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_address.R
\name{plot_address}
\alias{plot_address}
\title{Plot a given address/destination}
\usage{
plot_address(address, zoom = 10)
}
\arguments{
\item{address}{shall be a address/destination of class character}

\item{zoom}{shall be a integer of class character or numeric}
}
\value{
Returns a plot of the address/destination
}
\description{
Plot a given address/destination with diffrent zoom
}
\examples{
#plot_address("linkoping",13)
#plot_address("karlebo 101 hjorted",14)
}
